+++ src/server/devices/espresense/trilaterationPoint.ts
@@ -1,6 +1,25 @@
 import { Utils } from '../../services';
 
 export class TrilaterationPoint {
+  public static getPointsInRange(
+    a: TrilaterationPoint,
+    b: TrilaterationPoint,
+    roomName: string = '',
+  ): TrilaterationPoint[] {
+    const points: TrilaterationPoint[] = [];
+    if (a.x > b.x || a.y > b.y || a.z > b.z) {
+      throw new Error('Room Coordinates are not in order left-front-bottom to right-back-top');
+    }
+    for (let x = a.x; x <= b.x; x++) {
+      for (let y = a.y; y <= b.y; y++) {
+        for (let z = a.z; z <= b.z; z++) {
+          points.push(new TrilaterationPoint(x, y, z, roomName));
+        }
+      }
+    }
+    return points;
+  }
+
   public readonly coordinateName: string;
 
   constructor(